+++ src/dialogs/index.ts
@@ -1,11 +1,13 @@
 import { confirmDialog } from './confirm';
 import { inputDialog } from './input';
 import { createEntryDialog, updateEntryDialog } from './createEntry';
+import { saveChangesDialog } from './saveChanges';
 
 export class FCBDialog {
   static confirmDialog = confirmDialog;
   static inputDialog = inputDialog;
   static createEntryDialog = createEntryDialog;
   static updateEntryDialog = updateEntryDialog;
+  static saveChangesDialog = saveChangesDialog;
 
 }